% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{b3_ibra}
\alias{b3_ibra}
\title{Organizado o Ibra}
\usage{
b3_ibra(file)
}
\arguments{
\item{file}{<- "data-raw/nome_do_arquivo.csv".}
}
\value{
Retorna um data frame com os dados do Ibra na pasta "data/".
}
\description{
Organiza a planilha (csv) com o índice Ibra no site da B3.
}
\details{
Baixe a planilha no site http://www.b3.com.br/pt_br/market-data-e-indices/indices/indices-amplos/indice-brasil-amplo-ibra-composicao-da-carteira.htm e salve na pasta "data-raw/"
}
\examples{
file <- "data-raw/IBRADia_22-06-21.csv"
b3_ibra(file)

}
\author{
Max Lemes
}
